extract batch
